src/unified_catalog_client.py
"""
Unified Catalog Client
Provides interface for querying Microsoft Fabric Unified Catalog
including Data Products, Data Quality scores, and domain information.
"""

import logging
from typing import Dict, List, Any, Optional
from azure.identity import DefaultAzureCredential


logger = logging.getLogger(__name__)


class UnifiedCatalogClient:
    """Client for Microsoft Fabric Unified Catalog"""
    
    def __init__(self, workspace_id: str, credential=None):
        """
        Initialize Unified Catalog client
        
        Args:
            workspace_id: Fabric workspace ID
            credential: Azure credential (defaults to DefaultAzureCredential)
        """
        self.workspace_id = workspace_id
        self.credential = credential or DefaultAzureCredential()
    
    def get_data_products(self) -> List[Dict[str, Any]]:
        """
        Fetch all Data Products from Unified Catalog
        
        Returns:
            List of Data Product records with metadata
        """
        logger.info(f"Fetching Data Products from workspace: {self.workspace_id}")
        
        # This would use the Fabric SDK or REST API
        # Returns list of Data Products
        return []
    
    def get_data_product_assets(self, data_product_id: str) -> List[Dict[str, Any]]:
        """
        Get all assets in a specific Data Product
        
        Args:
            data_product_id: ID of the Data Product
            
        Returns:
            List of assets in the Data Product
        """
        logger.info(f"Fetching assets for Data Product: {data_product_id}")
        return []
    
    def get_data_quality_scores(self) -> List[Dict[str, Any]]:
        """
        Fetch all Data Quality scores from Unified Catalog
        
        Returns:
            List of data quality records with scores and metrics
        """
        logger.info("Fetching Data Quality scores from Unified Catalog")
        
        # Returns data quality information
        return []
    
    def get_domains(self) -> List[Dict[str, Any]]:
        """Fetch all Domains from Unified Catalog"""
        logger.info("Fetching Domains from Unified Catalog")
        return []
    
    def get_domain_content(self, domain_id: str) -> Dict[str, Any]:
        """Get all content (Data Products, assets) in a domain"""
        logger.info(f"Fetching content for domain: {domain_id}")
        return {}
    
    def extract_all_catalog_metadata(self) -> Dict[str, Any]:
        """
        Extract all metadata from Unified Catalog
        
        Returns:
            Dictionary containing all Unified Catalog metadata
        """
        metadata = {
            "domains": self.get_domains(),
            "data_products": self.get_data_products(),
            "data_quality": self.get_data_quality_scores(),
        }
        
        logger.info("Unified Catalog metadata extraction complete")
        return metadata


def get_unified_catalog_client(workspace_id: str, credential=None) -> UnifiedCatalogClient:
    """Factory function to create and configure a Unified Catalog client"""
    return UnifiedCatalogClient(workspace_id, credential)
