requirements.txt
# Core Dependencies
pyspark>=3.4.0
jupyter>=1.0.0
jupyterlab>=3.0.0

# Azure SDKs
azure-identity>=1.14.0
azure-purview-catalog>=1.0.0
azure-purview-administration>=1.0.0
azure-storage-blob>=12.18.0
azure-keyvault-secrets>=4.7.0

# Fabric / Power BI
fabric>=0.2.0
sempy>=0.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
polars>=0.19.0

# Configuration & Utils
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
tenacity>=8.2.0

# Logging & Monitoring
python-json-logger>=2.0.7

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.1
